Moving unicode testing out into its own file. cleaning up const/var usage. Using StreetName React component everywhere
