+++ app/views/ui/welcome.jade
@@ -18,10 +18,10 @@
         Welcome to Streetmix.
       p.
         This is
-        #[span.street-name#welcome-street-name]
+        #[span#welcome-street-name]
         <span id='welcome-avatar-creator'>
           made by
-          <span class='avatar' id='welcome-avatar' userId=''></span>
+          <span id='welcome-avatar'></span>
           <span id='welcome-creator'></span>
         </span>.
       p.important.
